Dockerize backend as well
