+++ .run/Run ALL.run.xml
@@ -1,10 +1,10 @@
 <component name="ProjectRunConfigurationManager">
-    <configuration default="false" name="Run DB" type="docker-deploy" factoryName="docker-compose.yml"
+    <configuration default="false" name="Run ALL" type="docker-deploy" factoryName="docker-compose.yml"
                    server-name="Docker">
         <deployment type="docker-compose.yml">
             <settings>
                 <option name="envFilePath" value=""/>
-                <option name="sourceFilePath" value="docker-compose-DB.yml"/>
+                <option name="sourceFilePath" value="docker-compose-ALL.yml"/>
             </settings>
         </deployment>
         <method v="2"/>
