+++ app/model/account.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 from pydantic import BaseModel
 
-from utils.as_form import as_form
+from app.utils.as_form import as_form
 
 class Account(BaseModel):
     account_uuid: str
