+++ app/model/general.py
