+++ app/server.py
@@ -1,8 +1,8 @@
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
-from core.settings import Settings
-from router import register_router
+from app.core.settings import Settings
+from app.router import register_router
 
 app = FastAPI(
     title="NFT API",
